restructure STREAM module
